refactor: layout & ssr
#144
