+++ packages/e-commerce-front/src/app/(main)/product/(list)/@products/loading.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import Grid from "@mui/material/Grid";
 import Skeleton from "@mui/material/Skeleton";
 import Stack from "@mui/material/Stack";
